Checked for cyclic dependency and refactored code as per gradle build
