+++ src/main/java/org/example/commands/AddDependencyCommand.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import java.util.Scanner;
 import org.example.exceptions.InvalidArgument;
 import org.example.model.Node;
+import org.example.validators.CyclicDependencyValidator;
 
 public class AddDependencyCommand implements CommandInterface {
 
@@ -17,6 +18,10 @@ public class AddDependencyCommand implements CommandInterface {
     String parentId = scanner.nextLine();
     String childId = scanner.nextLine();
 
+    // Check for possibility of any cycle
+    CyclicDependencyValidator cyclicDependencyValidator = new CyclicDependencyValidator();
+    cyclicDependencyValidator.checkForCycle(parentId, childId, nodeDependencies);
+
     // Check if both parent and child nodes exist in the map
     if (nodeDependencies.containsKey(parentId) && nodeDependencies.containsKey(childId)) {
 
@@ -28,7 +33,7 @@ public class AddDependencyCommand implements CommandInterface {
       nodeDependencies.get(parentId).setNodeChildren(children);
 
       // Update the child node: remove and add the parent ID to ensure it's not duplicated
-      ArrayList<String> parent =  nodeDependencies.get(childId).getNodeParents();
+      ArrayList<String> parent = nodeDependencies.get(childId).getNodeParents();
       parent.remove(parentId);
       parent.add(parentId);
       nodeDependencies.get(childId).setNodeParents(parent);
