--- src/main/java/org/example/validators/DependencyValidator.java
@@ -1,4 +1,0 @@
-package org.example.validators;
-
-public class DependencyValidator {
-}
